requirements.txt
# RealizeOS Core Dependencies

# LLM Clients
anthropic>=0.40.0
google-genai>=1.0.0

# Knowledge Base
pyyaml>=6.0

# Environment loading (auto-loads .env for non-Docker users)
python-dotenv>=1.0.0

# HTTP Client (web tools)
httpx>=0.27.0

# API Server (Full edition)
fastapi>=0.115.0
uvicorn>=0.32.0

# Database
# SQLite is built into Python — no extra deps

# Optional: Hybrid search (vector embeddings)
# fastembed>=0.4.0  # Uncomment for hybrid FTS5+vector search

# Optional: Telegram channel
# python-telegram-bot>=21.0  # Uncomment for Telegram integration

# Optional: Google Workspace tools
# google-api-python-client>=2.100.0
# google-auth-oauthlib>=1.2.0

# Optional: Web search
# trafilatura>=1.12.0  # URL content extraction

# Optional: Browser automation
# playwright>=1.48.0  # Headless browser
